extract preview component
