+++ course-react-ts-portfolio/jbook/src/index.tsx
@@ -4,11 +4,13 @@ import ReactDOM from 'react-dom';
 import { unpkgPathPlugin } from './plugins/unpkg-path-plugin';
 import { fetchPlugin } from './plugins/fetch-plugin';
 import CodeEditor from './components/code-editor';
+import Preview from './components/preview';
 
 const App = () => {
   const ref = useRef<any>();
-  const iFrame = useRef<any>();
+
   const [input, setInput] = useState('');
+  const [code, setCode] = useState('');
 
   const startService = async () => {
     ref.current = await esBuild.startService({
@@ -26,9 +28,6 @@ const App = () => {
       return;
     }
 
-    // reset content of iframe
-    iFrame.current.srcdoc = html;
-
     const result = await ref.current.build({
       entryPoints: ['index.js'],
       bundle: true,
@@ -40,30 +39,8 @@ const App = () => {
       },
     });
 
-    // Post bundled code as message to iFrame
-    iFrame.current.contentWindow.postMessage(result.outputFiles[0].text, '*');
+    setCode(result.outputFiles[0].text);
   };
-
-  // place code inside script tag of iframe html document
-  const html = `
-    <html>
-      <head></head>
-      <body>
-        <div id="root"></div>
-        <script>
-          window.addEventListener('message', (event) => {
-            try {
-              eval(event.data);
-            } catch(err) {
-              const root = document.querySelector('#root');
-              root.innerHTML = '<div style="color: red"><h4>Runtime Error</h4>' + err + '</div>';
-              console.error(err);
-            }
-          }, false);
-        </script>
-      </body>
-    </html>
-  `;
 
   return (
     <div>
@@ -71,19 +48,10 @@ const App = () => {
         initialValue='const a = 1;'
         onChange={(value) => setInput(value)}
       />
-      <textarea
-        value={input}
-        onChange={(e) => setInput(e.target.value)}
-      ></textarea>
       <div>
         <button onClick={onClick}>Submit</button>
       </div>
-      <iframe
-        ref={iFrame}
-        sandbox='allow-scripts'
-        srcDoc={html}
-        title='preview'
-      ></iframe>
+      <Preview code={code} />
     </div>
   );
 };
